requirements.txt
aenum==2.1.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==1.0.2
Flask-WTF==0.14.2
futures==3.2.0
grpcio==1.15.0
grpcio-tools==1.15.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
maprdb-python-client==1.0.0
MarkupSafe==1.0
ojai-python-api==1.0
pkginfo==1.4.2
prettytable==0.7.2
protobuf==3.6.1
python-dateutil==2.7.3
requests==2.18.4
requests-toolbelt==0.8.0
retrying==1.3.3
six==1.11.0
tqdm==4.23.3
twine==1.11.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.2.1
yolk3k==0.9
